chore(rig-vertexai): move inline test modules to sibling files
